ios+android
안드로이드 Ios 애뮬 확인
